refactor: 대댓글 수
