+++ be/src/main/java/com/foodymoody/be/comment/application/CommentMapper.java
@@ -8,10 +8,10 @@ import com.foodymoody.be.comment.application.dto.response.MemberReplySummaryResp
 import com.foodymoody.be.comment.application.dto.response.MemberSummaryResponse;
 import com.foodymoody.be.comment.domain.entity.Comment;
 import com.foodymoody.be.comment.domain.entity.Reply;
-import com.foodymoody.be.comment.domain.entity.ReplyId;
 import com.foodymoody.be.common.util.ids.CommentId;
 import com.foodymoody.be.common.util.ids.FeedId;
 import com.foodymoody.be.common.util.ids.MemberId;
+import com.foodymoody.be.common.util.ids.ReplyId;
 import java.time.LocalDateTime;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
