+++ be/src/main/java/com/foodymoody/be/comment/domain/repository/ReplyRepository.java
@@ -2,9 +2,9 @@ package com.foodymoody.be.comment.domain.repository;
 
 import com.foodymoody.be.comment.application.dto.response.MemberReplySummary;
 import com.foodymoody.be.comment.domain.entity.Reply;
-import com.foodymoody.be.comment.domain.entity.ReplyId;
 import com.foodymoody.be.common.util.ids.CommentId;
 import com.foodymoody.be.common.util.ids.MemberId;
+import com.foodymoody.be.common.util.ids.ReplyId;
 import java.util.Optional;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
