+++ be/src/main/java/com/foodymoody/be/comment/infra/persistence/ReplyRepositoryImpl.java
@@ -2,11 +2,11 @@ package com.foodymoody.be.comment.infra.persistence;
 
 import com.foodymoody.be.comment.application.dto.response.MemberReplySummary;
 import com.foodymoody.be.comment.domain.entity.Reply;
-import com.foodymoody.be.comment.domain.entity.ReplyId;
 import com.foodymoody.be.comment.domain.repository.ReplyRepository;
 import com.foodymoody.be.comment.infra.persistence.jpa.ReplyJpaRepository;
 import com.foodymoody.be.common.util.ids.CommentId;
 import com.foodymoody.be.common.util.ids.MemberId;
+import com.foodymoody.be.common.util.ids.ReplyId;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
