+++ be/src/main/java/com/foodymoody/be/reply_heart/infra/persistence/ReplyHeartRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.foodymoody.be.reply_heart.infra.persistence;
 
-import com.foodymoody.be.comment.domain.entity.ReplyId;
+import com.foodymoody.be.common.util.ids.ReplyId;
 import com.foodymoody.be.reply_heart.domain.ReplyHeart;
 import com.foodymoody.be.reply_heart.domain.ReplyHeartRepository;
 import com.foodymoody.be.reply_heart.infra.persistence.jpa.ReplyHeartJpaRepository;
