+++ be/src/main/java/com/foodymoody/be/reply_heart/infra/usecase/ReplyHeartAddedEvent.java
@@ -1,10 +1,10 @@
 package com.foodymoody.be.reply_heart.infra.usecase;
 
-import com.foodymoody.be.comment.domain.entity.ReplyId;
 import com.foodymoody.be.common.event.NotificationType;
 import com.foodymoody.be.common.util.ids.CommentId;
 import com.foodymoody.be.common.util.ids.FeedId;
 import com.foodymoody.be.common.util.ids.MemberId;
+import com.foodymoody.be.common.util.ids.ReplyId;
 import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
