+++ be/src/test/java/com/foodymoody/be/acceptance/reply_heart/ReplyFeedHeartAcceptanceTest.java
@@ -1,9 +1,9 @@
 package com.foodymoody.be.acceptance.reply_heart;
 
-import static com.foodymoody.be.acceptance.comment.CommentSteps.댓글에_댓글을_등록한다;
-import static com.foodymoody.be.acceptance.comment.CommentSteps.댓글의_댓글을_조회한다;
 import static com.foodymoody.be.acceptance.comment.CommentSteps.피드에_댓글을_등록하고_아이디를_받는다;
 import static com.foodymoody.be.acceptance.feed.FeedSteps.피드를_등록하고_아이디를_받는다;
+import static com.foodymoody.be.acceptance.reply.ReplySteps.댓글에_댓글을_등록한다;
+import static com.foodymoody.be.acceptance.reply.ReplySteps.댓글의_댓글을_조회한다;
 import static com.foodymoody.be.acceptance.reply_heart.ReplyHeartSteps.좋아요를_누린다;
 import static com.foodymoody.be.acceptance.reply_heart.ReplyHeartSteps.좋아요를_취소한다;
 import static org.assertj.core.api.Assertions.assertThat;
